// src/app/lib/enum.ts
export enum InputState {
  valid = "valid",
  invalid = "invalid",
  idle = "idle",
  none = "none",
};

export enum InputType {
  email = "email",
  phone = "phone"
};

export enum InputStyleType {
  valid = "input-field-valid",
  invalid = "input-field-invalid",
  fadeIn = "input-field-fade-in",
  fadeOut = "input-field-fade-out",
  default = "input-field",
};